Fix imports after renaming modules
