+++ tfep/tests/nn/dynamics/test_egnn.py
@@ -6,7 +6,7 @@
 # =============================================================================
 
 """
-Test classes and functions in tfep.nn.modules.egnn.
+Test classes and functions in tfep.nn.dynamics.egnn.
 """
 
 
@@ -19,7 +19,7 @@ import pytest
 import scipy.spatial.transform
 import torch
 
-from tfep.nn.graph.egnn import EGNNDynamics, _EGLayer, _unsorted_segment_sum
+from tfep.nn.dynamics.egnn import EGNNDynamics, _EGLayer, _unsorted_segment_sum
 from tfep.utils.geometry import batchwise_rotate
 from tfep.utils.misc import atom_to_flattened, flattened_to_atom
